+++ examples/shared/jest/index.ts
@@ -1,0 +1,3 @@
+import { makeChart } from "./src/chart";
+
+makeChart();